backend/requirements.txt
fastapi>=0.109.1
uvicorn==0.27.0
pydantic==2.11.7
pydantic-settings==2.9.1
langchain>=0.0.300
langchain-openai==0.3.23
langchain-core==0.3.65
langgraph==0.4.8
tavily-python==0.5.4
python-dotenv==1.0.0
typing_extensions==4.13.2
python-jose